+++ script.js
@@ -8,17 +8,18 @@ const taskInput = document.getElementById("taskInput");
 const taskList = document.getElementById("taskList");
 const addTaskBtn = document.getElementById("addTaskBtn");
 
+// TODO
 // Variable used to temporarily store updated task text
 let updateText
 
 // Get saved tasks from localStorage
-let storedTasks = JSON.parse(localStorage.getItem("storedTasks"))
-console.log("storedTasks ", storedTasks)
+let storedTasks = GetLocalStorageTasksValue()
 
 // ==========================================
 // EVENT LISTENERS
 // ==========================================
 
+// TODO
 // Detect when Enter key is pressed in the input field and automatically click the add/update button
 taskInput.addEventListener("keypress", function (e) {
     if (e.key === "Enter") {
@@ -32,8 +33,8 @@ taskInput.addEventListener("keypress", function (e) {
 // INITIAL PAGE LOAD
 // ==========================================
 
-// Display all saved tasks when the page loads
-ReadToDoItems();
+// Display all saved tasks from the local storage in the UI when the page loads
+CreateTaskListUIFromLocalStorage();
 
 
 
@@ -41,25 +42,47 @@ ReadToDoItems();
 // FUNCTIONS
 // ==========================================
 
-// Display all todo items from storedTasks on the webpage
-// Add a line through style to completed items and show them in the list
-function ReadToDoItems() {
-    storedTasks.forEach((element) => {
-        let li = document.createElement("li");
-        let style = "";
+/**
+ * Returns the list of stored tasks from localStorage.
+ * Returns an empty array if no tasks are found.
+ *
+ * @returns {List} List of task objects
+ */
+function GetLocalStorageTasksValue() {
+    // Get the local storage data of the saved tasks
+    let storedTasks = JSON.parse(localStorage.getItem("storedTasks"))
 
-        // If this task is marked as completed in localStorage, display it crossed out in the UI
-        if (element.isComplete) {
-            style = "style='text-decoration: line-through'";
-        }
+    // If the local storage doesnt contain any tasks, create an empty list
+    if (storedTasks === null) {
+        storedTasks = [];
+    }
 
-        const todoItems = `<div ${style} ondblclick="CompleteToDoItems(this)">
-            ${element.item}${style === ""
-                ? '<img class="edit todo-controls" onclick="UpdateToDoItems(this)" src="/images/pencil.png" />'
-                : ""
-            }<img class="delete todo-controls" onclick="DeleteToDoItems(this)" src="/images/delete.png" /></div>`;
-        li.innerHTML = todoItems;
-        taskList.appendChild(li);
+    console.log("storedTasks init", storedTasks)
+    return storedTasks;
+}
+
+/**
+ * Save the current value of the global variable "storedTasks" inside the local storage browser
+ * Have to call this function everytime the global variable storedTasks change to have the local storage browser up to date
+ */
+function SetLocalStorageTasksValue() {
+    localStorage.setItem("storedTasks", JSON.stringify(storedTasks));
+}
+
+/**
+ * Creating the list of task in the UI from the local storage
+ */
+function CreateTaskListUIFromLocalStorage() {
+    // Going through each saved Task object in the local storage
+    storedTasks.forEach((savedTask) => {
+        // Create a new task item (li) with the task text from the saved task from the local storage
+        const newTaskItem = CreateTaskElement(savedTask.taskText)
+
+        // Add the newTaskItem inside the taskList
+        taskList.appendChild(newTaskItem);
+
+        // Apply the UI completion on the task element
+        ApplyTaskCompletionUI(newTaskItem.querySelector(".task-text"), savedTask.isComplete);
     });
 }
 
@@ -70,23 +93,24 @@ function ReadToDoItems() {
  * - Clears the input field
  */
 function OnAddTaskClick() {
+    // Create a copy of the global variable "taskInput.value" to use it instead of using the global variable too much
+    const taskText = taskInput.value;
+
     // Check if the task input field is empty
     // If it is, display an alert message and stop the function
-    if (taskInput.value === "") {
+    if (taskText === "") {
         return alert("Please Enter your todo text!");
     }
 
-    const taskText = taskInput.value;
-
     // Create a new task item (li) and define its HTML content for the task text and action buttons (Edit/Delete)
     // Clicking the task text marks the task as completed by toggling a line-through style
-    let newTaskItem = createTaskElement(taskText)
-    
+    const newTaskItem = CreateTaskElement(taskText)
+
     // Add the newTaskItem inside the taskList
     taskList.appendChild(newTaskItem);
-  
+
     // Call the function to create and save the new task in the local storage
-    saveNewTaskInLocalStorage(taskText)
+    SaveNewTaskInLocalStorage(taskText)
 
     // Reset the input field so the user can type a new task
     taskInput.value = "";
@@ -101,8 +125,8 @@ function OnAddTaskClick() {
  */
 function CreateTaskElement(taskText) {
     let newTaskItem = document.createElement("li");
-    newTaskItem.innerHTML = `<div ondblclick="CompleteTask(this)">${taskText}</div>
-        <div>
+    newTaskItem.innerHTML = `<div class="task-text" ondblclick="CompleteTask(this)">${taskText}</div>
+        <div class="task-actions">
             <img class="edit todo-controls" onclick="UpdateToDoItems(this)" src="images/pencil.png" />
             <img class="delete todo-controls" onclick="DeleteToDoItems(this)" src="images/delete.png"/>
         </div>`;
@@ -116,7 +140,7 @@ function CreateTaskElement(taskText) {
  *
  * @param {string} taskText - The text of the new task
  */
-function SaveNewTaskInLocalStorage(taskText){
+function SaveNewTaskInLocalStorage(taskText) {
     // Create a new task object for the local storage
     let newTaskData = {
         taskText: taskText,
@@ -127,45 +151,63 @@ function SaveNewTaskInLocalStorage(taskText){
     storedTasks.push(newTaskData);
 
     // Save the current tasks list to localStorage
-    setLocalStorage();
+    SetLocalStorageTasksValue();
+
+    console.log("storedTasks saved", storedTasks)
 }
 
 /**
- * CompleteTaskItems Function
- * Toggles the completion state of a todo item on double-click.
- * Applies or removes strikethrough styling and shows/hides the edit button accordingly.
+ * Toggles the completion state of a task
  *
- * @param {HTMLElement} e - The clicked todo element
+ * @param {HTMLElement} taskElement - The task element
  */
-function CompleteTask(e) {
-    // Get the edit button inside the todo item container
-    const editBtn = e.parentElement.querySelector("img.edit");
+function CompleteTask(taskElement) {
+    // Search the task to complete by his text from the local storage list tasks
+    storedTasks.forEach((savedTask) => {
+        if (taskElement.parentElement.querySelector("div").innerText.trim() === savedTask.taskText) {
+            // Switch the status of the task
+            if (savedTask.isComplete == true) {
+                savedTask.isComplete = false
+            } else {
+                savedTask.isComplete = true
+            }
 
-    // Change the status of the todo item
-    if (e.isComplete == true) {
-        e.isComplete = false
-    } else {
-        e.isComplete = true
-    }
+            // Applies or removes strike through styling and shows/hides the edit button accordingly
+            ApplyTaskCompletionUI(taskElement, savedTask.isComplete)
+        }
+    });
 
-    // Applies or removes strikethrough styling and shows/hides the edit button accordingly.
-    if (e.isComplete == true) {
-        e.style.textDecoration = "none";
-        if (editBtn) editBtn.style.display = "inline-block";
-    } else {
-        e.style.textDecoration = "line-through";
-        if (editBtn) editBtn.style.display = "none";
-    }
-
-    // TODO: Save the current completion state of each task in localStorage
-    //storedTasks.forEach((element) => {
-    //    if (e.parentElement.querySelector("div").innerText.trim() === element.item) {
-    //        element.isComplete = true;
-    //    }
-    //});
+    // Save the current tasks list to localStorage
+    SetLocalStorageTasksValue();
+    console.log("storedTasks after complete", storedTasks)
 }
 
+/**
+ * Updates the task UI based on its completion state
+ * Applies strikethrough styling and toggles the edit button
+ *
+ * @param {HTMLElement} taskElement - The task element
+ * @param {boolean} isComplete - Whether the task is completed
+ */
+function ApplyTaskCompletionUI(taskElement, isComplete) {
+    // Get the edit button inside the task element
+    const li = taskElement.closest("li");
+    const editBtn = li.querySelector("img.edit");
+    console.log("editBtn", editBtn)
 
+    // Applies or removes strike through styling and shows/hides the edit button
+    if (isComplete == false) {
+        taskElement.style.textDecoration = "none";
+        editBtn.style.display = "inline-block";
+    } else {
+        taskElement.style.textDecoration = "line-through";
+        editBtn.style.display = "none";
+    }
+
+    console.log("editBtn after applis style", editBtn)
+}
+
+// TODO
 function UpdateOnSelectionItems() {
 
 
@@ -175,11 +217,11 @@ function UpdateOnSelectionItems() {
         }
     });
 
-    setLocalStorage();
+    SetLocalStorageTasksValue();
     taskInput.value = "";
 }
 
-
+// TODO
 //Change the add button into an update button when editing a todo item
 function UpdateToDoItems(e) {
     updateText = e.parentElement.parentElement.querySelector("div");
@@ -190,6 +232,7 @@ function UpdateToDoItems(e) {
     addTaskBtn.setAttribute("src", "/images/refresh.png");
 }
 
+// TODO
 function DeleteToDoItems(e) {
     let deleteValue =
         e.parentElement.parentElement.querySelector("div").innerText;
@@ -202,10 +245,6 @@ function DeleteToDoItems(e) {
                 storedTasks.splice(element, 1);
             }
         });
-        setLocalStorage();
+        SetLocalStorageTasksValue();
     }
-}
-
-function setLocalStorage() {
-    localStorage.setItem("storedTasks", JSON.stringify(storedTasks));
 }